feat/OPS-61: added auto documentation for haiku.py, main.py and store_manager.py
At the moment it works by converting senju into a python package. This causes the html docu to start with a link 'senju package' which lists all three moduls and their documentation. I will try now to make it so all three moduls have a docu link of their own.

The method foobar in each modul is for testing purposes and should be deleted once full functionality is confirmed.